requirements.txt
tensorflow-cpu==2.15.0
numpy==1.26.4
Pillow>=10.0.0
opencv-python-headless>=4.8.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
streamlit>=1.31.0
tqdm>=4.66.0
